perf: optimise HTTP server hot path
- Replace Server::_connections vector with unordered_map for O(1) lookup
- Single-pass response header assembly with pre-computed reserve
- Add NullSharedMutex and LocalSignal for lock-free signal dispatch on
  single-threaded event loops (libuv guarantees single-thread per loop)
- Switch all net/http signals to LocalSignal, eliminating mutex
  acquisition and slots vector copy on every emit
- Add move semantics for parser header flush into NVCollection
- Single-receiver fast path in SocketAdapter dispatch (skip loop for the
  common case of exactly 1 receiver per connection)
- Fix LogStream stub destructor for ENABLE_LOGGING=OFF builds
- Fix basetests compile with logging disabled
- Fix httpclient StatusCode enum streaming

+++ src/http/samples/httpclient/httpclient.cpp
@@ -39,7 +39,7 @@ int main(int argc, char** argv)
     auto conn = http::createConnectionT<http::ClientConnection>(URL(url));
 
     conn->Headers += [](http::Response& res) {
-        std::cout << "Status: " << res.getStatus() << " " << res.getReason() << std::endl;
+        std::cout << "Status: " << static_cast<int>(res.getStatus()) << " " << res.getReason() << std::endl;
         std::cout << "Headers:" << std::endl;
         for (const auto& [name, value] : res) {
             std::cout << "  " << name << ": " << value << std::endl;
